Make flags BOOLEAN and widen Shamir fields
Change migration columns is_admin, tainted and revertible to BOOLEAN.
Update Shamir.total_shares and threshold from i16 to i32 in core and db
representations. Remove explicit generic type on sqlx::query_scalar.

+++ decodering-core/src/repository.rs
@@ -246,8 +246,8 @@ pub trait AuditRepository: Send {
 #[derive(Debug, Clone)]
 pub struct Shamir {
     pub id: i64,
-    pub total_shares: i16,
-    pub threshold: i16,
+    pub total_shares: i32,
+    pub threshold: i32,
     pub validation_hash: Vec<u8>,
     pub created_at: i64,
 }
